feat: move app-store page to RSC
