+++ src/app/(dashboard)/app-store/components/AppStoreTable/index.ts
@@ -1,0 +1,1 @@
+export { AppStoreTable } from './AppStoreTable';
