+++ src/app/(dashboard)/app-store/components/AppStoreTableActions/AppStoreTableActions.module.scss
