+++ src/app/(dashboard)/app-store/components/AppStoreTile/index.ts
@@ -1,0 +1,1 @@
+export { AppStoreTile } from './AppStoreTile';
