+++ src/app/(dashboard)/app-store/components/CategorySelector/index.ts
@@ -1,0 +1,1 @@
+export { CategorySelector } from './CategorySelector';
