--- src/client/modules/AppStore/components/AppStoreTable/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppStoreTable';
