--- src/client/modules/AppStore/components/AppStoreTile/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './AppStoreTile';
