--- src/client/modules/AppStore/components/CategorySelector/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CategorySelector';
