--- src/client/modules/AppStore/containers/AppStoreContainer/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppStoreContainer';
