--- src/client/modules/AppStore/pages/AppStorePage/index.ts
@@ -1,1 +1,0 @@
-export { AppStorePage } from './AppStorePage';
